use context to control backup flow
