+++ br/pkg/backup/push.go
@@ -4,21 +4,23 @@ package backup
 
 import (
 	"context"
+	"io"
+	"os"
+	"time"
 
+	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
 	backuppb "github.com/pingcap/kvproto/pkg/brpb"
 	"github.com/pingcap/kvproto/pkg/errorpb"
+	"github.com/pingcap/kvproto/pkg/kvrpcpb"
+	"github.com/pingcap/log"
 	"github.com/pingcap/tidb/br/pkg/logutil"
+	"github.com/pingcap/tidb/br/pkg/rtree"
 	"github.com/pingcap/tidb/br/pkg/utils"
 	"go.uber.org/zap"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
-
-// pushDown wraps a backup task.
-type pushDown struct {
-	mgr    ClientMgr
-	respCh chan responseAndStore
-	errCh  chan error
-}
 
 type responseAndStore struct {
 	Resp    *backuppb.BackupResponse
@@ -31,6 +33,77 @@ func (r responseAndStore) GetResponse() *backuppb.BackupResponse {
 
 func (r responseAndStore) GetStoreID() uint64 {
 	return r.StoreID
+}
+
+func doSendBackup(
+	ctx context.Context,
+	client backuppb.BackupClient,
+	req backuppb.BackupRequest,
+	respFn func(*backuppb.BackupResponse) error,
+) error {
+	failpoint.Inject("hint-backup-start", func(v failpoint.Value) {
+		logutil.CL(ctx).Info("failpoint hint-backup-start injected, " +
+			"process will notify the shell.")
+		if sigFile, ok := v.(string); ok {
+			file, err := os.Create(sigFile)
+			if err != nil {
+				log.Warn("failed to create file for notifying, skipping notify", zap.Error(err))
+			}
+			if file != nil {
+				file.Close()
+			}
+		}
+		time.Sleep(3 * time.Second)
+	})
+	bCli, err := client.Backup(ctx, &req)
+	failpoint.Inject("reset-retryable-error", func(val failpoint.Value) {
+		switch val.(string) {
+		case "Unavaiable":
+			{
+				logutil.CL(ctx).Debug("failpoint reset-retryable-error unavailable injected.")
+				err = status.Error(codes.Unavailable, "Unavailable error")
+			}
+		case "Internal":
+			{
+				logutil.CL(ctx).Debug("failpoint reset-retryable-error internal injected.")
+				err = status.Error(codes.Internal, "Internal error")
+			}
+		}
+	})
+	failpoint.Inject("reset-not-retryable-error", func(val failpoint.Value) {
+		if val.(bool) {
+			logutil.CL(ctx).Debug("failpoint reset-not-retryable-error injected.")
+			err = status.Error(codes.Unknown, "Your server was haunted hence doesn't work, meow :3")
+		}
+	})
+	if err != nil {
+		return err
+	}
+	defer func() {
+		_ = bCli.CloseSend()
+	}()
+
+	for {
+		resp, err := bCli.Recv()
+		if err != nil {
+			if errors.Cause(err) == io.EOF { // nolint:errorlint
+				logutil.CL(ctx).Debug("backup streaming finish",
+					logutil.Key("backup-start-key", req.GetStartKey()),
+					logutil.Key("backup-end-key", req.GetEndKey()))
+				return nil
+			}
+			return err
+		}
+		// TODO: handle errors in the resp.
+		logutil.CL(ctx).Debug("range backed up",
+			logutil.Key("small-range-start-key", resp.GetStartKey()),
+			logutil.Key("small-range-end-key", resp.GetEndKey()),
+			zap.Int("api-version", int(resp.ApiVersion)))
+		err = respFn(resp)
+		if err != nil {
+			return errors.Trace(err)
+		}
+	}
 }
 
 func startStoreBackup(
@@ -48,7 +121,7 @@ func startStoreBackup(
 		retry := -1
 		return utils.WithRetry(ctx, func() error {
 			retry += 1
-			logutil.CL(ctx).Info("try backup", zap.Int("retry time", retry))
+			logutil.CL(ctx).Info("try backup", zap.Uint64("storeID", storeID), zap.Int("retry time", retry))
 			// Send backup request to the store.
 			// handle the backup response or internal error here.
 			// handle the store error(reboot or network partition) outside.
@@ -99,4 +172,15 @@ func startStoreBackup(
 			})
 		}, utils.NewBackupSSTBackoffer())
 	}
+}
+
+func getBackupRanges(ranges []rtree.Range) []*kvrpcpb.KeyRange {
+	requestRanges := make([]*kvrpcpb.KeyRange, 0, len(ranges))
+	for _, r := range ranges {
+		requestRanges = append(requestRanges, &kvrpcpb.KeyRange{
+			StartKey: r.StartKey,
+			EndKey:   r.EndKey,
+		})
+	}
+	return requestRanges
 }
